Changes to Enemy AI and PlayerHealth

+++ Assets/IDamageable.cs
@@ -4,5 +4,5 @@ public interface IDamageable
     //this is not a class it's an interface, it's a contract that any class that implements it must follow
     // this means that any class that implements this interface must have a method called ReceiveDamage
 {
-    void ReceiveDamage(float damageTaken, GameObject source);
+    void ReceiveDamage(int damageTaken, GameObject source);
 }
